2021.03.03_algorithm

+++ jongsik_won/13305.cpp
@@ -1,20 +1,20 @@
 #include <cstdio>
 
-int len[100000];
-int cost[100000];
+long long len[100000];
+long long cost[100000];
 
 int main()
 {
-    int ans=0;
+    long long ans=0;
     int count=0;
     scanf("%d",&count);
     for(int i=0;i<count-1;i++)
     {
-        scanf("%d",&len[i]);
+        scanf("%lld",&len[i]);
     }
     for(int i=0;i<count;i++)
     {
-        scanf("%d",&cost[i]);
+        scanf("%lld",&cost[i]);
     }
     ans=cost[0]*len[0];
     for(int i=1;i<count-1;i++)
@@ -25,5 +25,5 @@ int main()
         }
         ans+=cost[i]*len[i];
     }
-    printf("%d",ans);
+    printf("%lld",ans);
 }
